Switch from moxios to axios-mock-adapter.
I had assumed that moxios was well-supported since it's part
of the axios project on GitHub, but it's actually buggy and
unmaintained (https://github.com/axios/moxios/issues/53). In
particular, I had tons of timing issues and problems getting
it to send proper responses while writing tests for the
Import view.