Downgrade to @vue/test-utils@1.0.0-beta.29 from .31.
Downgrade to the latest version of vue-test-utils that
includes sync mode since it seems to work around the
v-treeview issue that forced me to disable the Ticks view's
initial-route test. Sync mode breaks the Import view's
tests, so I need to explicitly disable it there.

Sync mode was apparently removed in 1.0.0-beta.30
(https://github.com/vuejs/vue-test-utils/issues/1130), so
this is probably going to be a big "screw you" to my future
self. There's another bug introduced in .30 that I want to
avoid (https://github.com/vuejs/vue-test-utils/issues/1407),
so it's serendipitous for now, though.